requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.1.0
bcrypt==4.2.1
black==24.8.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
dnspython==2.7.0
docx2txt==0.8
email_validator==2.2.0
fastembed==0.5.1
filelock==3.17.0
filetype==1.2.0
flake8==7.1.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.47
langchain-google-genai==2.1.0
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.6
langsmith==0.3.10
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.15
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
onnxruntime==1.20.1
orjson==3.10.15
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psycopg2-binary==2.9.10
py_rust_stemmers==0.1.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.3
pyreadline3==3.5.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
qdrant-client==1.13.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
setuptools==75.8.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
win32_setctime==1.2.0
yarl==1.18.3
zstandard==0.23.0
